+++ changelog.md
@@ -1,2 +1,5 @@
+### Trash Cans 1.0.18a
+- Account for breaking changes in Mekanism
+
 ### Trash Cans 1.0.18
 - Initial release of Trash Cans for Minecraft 1.21